akalia_dev/requirements.txt
gunicorn
flask
bcrypt
celery
click-didyoumean
cryptography
exceptiongroup
Flask
Flask-Bootstrap
Flask-Limiter
Flask-Login
Flask-WTF
futures
httpcore
httpx
idna
jsonify
limits
loguru
psutil
shutils
python-dotenv
python-telegram-bot
requests
traceroute
typing_extensions
urllib3
wrapt
WTForms
flask_sqlalchemy
eventlet
flask_socketio